+++ NVEncCore/rgy_caption.h
@@ -240,7 +240,7 @@ public:
     Caption2Ass();
     virtual ~Caption2Ass();
     RGY_ERR init(std::shared_ptr<RGYLog> pLog, C2AFormat format);
-    RGY_ERR proc(const uint8_t *data, const int64_t data_size, std::vector<AVPacket>& subList);
+    RGY_ERR proc(const uint8_t *data, const size_t data_size, std::vector<AVPacket>& subList);
     void close();
     bool enabled() const { return !!m_dll; };
     void setVidFirstKeyPts(int64_t pts) {
@@ -253,7 +253,7 @@ public:
     C2AFormat format() const { return m_format; }
 
     //入力データがtsかどうかの判定
-    bool isTS(const uint8_t *data, const int64_t data_size) const;
+    bool isTS(const uint8_t *data, const size_t data_size) const;
 
     //出力解像度の設定
     void setOutputResolution(int w, int h, int sar_x, int sar_y);
